Move tclk system test to own file
